Refactor app to mattermost.android.js
